requirements.txt
streamlit
google-generativeai
python-docx
PyMuPDF
requests
streamlit-lottie
wordcloud
matplotlib
fpdf
python-dotenv
streamlit-option-menu
